Use a hack to enable DNS resolution of pods
Pods need a stable address to each other to register with
repmgr. However, pods get a new IP address each time they
get restarted. Register a k8s Headless Service for each
pod, using a selector to ensure connections are routed
correctly, allowing us to use the DNS name of the
service to connect to PostgreSQL on the pods.